errors/base_test.go
package errors

import (
	"io"
	"testing"
)

func TestBase(t *testing.T) {
	err := newBase("error")
	s := err.Error()
	expected := "error"
	if s != expected {
		t.Fatalf("unexpected message: got %q, want %q", s, expected)
	}
}

func TestNew(t *testing.T) {
	err := New("error")
	s := err.Error()
	expected := "error"
	if s != expected {
		t.Fatalf("unexpected message: got %q, want %q", s, expected)
	}
	sfs := StackFrames(err)
	if len(sfs) != 1 {
		t.Fatalf("unexpected length: got %d, want %d", len(sfs), 1)
	}
}

func TestNewf(t *testing.T) {
	err := Newf("%s", "error")
	s := err.Error()
	expected := "error"
	if s != expected {
		t.Fatalf("unexpected message: got %q, want %q", s, expected)
	}
	sfs := StackFrames(err)
	if len(sfs) != 1 {
		t.Fatalf("unexpected length: got %d, want %d", len(sfs), 1)
	}
}

func TestIs(t *testing.T) {
	err := io.EOF
	err = Wrap(err, "test")
	ok := Is(err, io.EOF)
	if !ok {
		t.Fatal("not ok")
	}
}
